+++ CHANGELOG.md
@@ -8,6 +8,12 @@
 
 * Swift 1.x is no longer supported.
 
+* `--templates-directory` and `--assets-directory` have been deprecated in favor
+  of `--theme`. Specify either 'apple' (default) or the path to your mustache
+  templates and other assets for a custom theme.  
+  [JP Simard](https://github.com/jpsim)
+  [#130](https://github.com/realm/jazzy/issues/130)
+
 ##### Enhancements
 
 * Add `--sdk [iphone|watch|appletv][os|simulator]|macosx` option for Objective-C